+++ packages/eas-cli/src/commands/channel/view.ts
@@ -1,141 +1,15 @@
-import { Flags } from '@oclif/core';
 import assert from 'assert';
-import chalk from 'chalk';
-import Table from 'cli-table3';
 
+import {
+  listAndRenderBranchesAndUpdatesOnChannelAsync,
+  selectChannelOnAppAsync,
+} from '../../channel/queries';
 import EasCommand from '../../commandUtils/EasCommand';
-import { ChannelQuery } from '../../graphql/queries/ChannelQuery';
-import Log from '../../log';
+import { EasNonInteractiveAndJsonFlags } from '../../commandUtils/flags';
+import { EasPaginatedQueryFlags, getPaginatedQueryOptions } from '../../commandUtils/pagination';
 import { getExpoConfig } from '../../project/expoConfig';
 import { findProjectRootAsync, getProjectIdAsync } from '../../project/projectUtils';
-import { promptAsync } from '../../prompts';
-import {
-  FormatUpdateParameter,
-  UPDATE_COLUMNS,
-  formatUpdateMessage,
-  getPlatformsForGroup,
-} from '../../update/utils';
-import formatFields from '../../utils/formatFields';
-import { enableJsonOutput, printJsonOnlyOutput } from '../../utils/json';
-
-export type BranchMapping = {
-  version: number;
-  data: {
-    branchId: string;
-    branchMappingLogic: {
-      operand: number;
-      clientKey: string;
-      branchMappingOperator: string;
-    } & string;
-  }[];
-};
-
-/**
- * Get the branch mapping and determine whether it is a rollout.
- * Ensure that the branch mapping is properly formatted.
- */
-export function getBranchMapping(branchMappingString?: string): {
-  branchMapping: BranchMapping;
-  isRollout: boolean;
-  rolloutPercent?: number;
-} {
-  if (!branchMappingString) {
-    throw new Error('Missing branch mapping.');
-  }
-  let branchMapping: BranchMapping;
-  try {
-    branchMapping = JSON.parse(branchMappingString);
-  } catch {
-    throw new Error(`Could not parse branchMapping string into a JSON: "${branchMappingString}"`);
-  }
-  assert(branchMapping, 'Branch Mapping must be defined.');
-
-  if (branchMapping.version !== 0) {
-    throw new Error('Branch mapping must be version 0.');
-  }
-
-  const isRollout = branchMapping.data.length === 2;
-  const rolloutPercent = branchMapping.data[0].branchMappingLogic.operand;
-
-  switch (branchMapping.data.length) {
-    case 0:
-      break;
-    case 1:
-      if (branchMapping.data[0].branchMappingLogic !== 'true') {
-        throw new Error('Branch mapping logic for a single branch must be "true"');
-      }
-      break;
-    case 2:
-      if (branchMapping.data[0].branchMappingLogic.clientKey !== 'rolloutToken') {
-        throw new Error('Client key of initial branch mapping must be "rolloutToken"');
-      }
-      if (branchMapping.data[0].branchMappingLogic.branchMappingOperator !== 'hash_lt') {
-        throw new Error('Branch mapping operator of initial branch mapping must be "hash_lt"');
-      }
-      if (rolloutPercent == null) {
-        throw new Error('Branch mapping is missing a "rolloutPercent"');
-      }
-      if (branchMapping.data[1].branchMappingLogic !== 'true') {
-        throw new Error('Branch mapping logic for a the second branch of a rollout must be "true"');
-      }
-      break;
-    default:
-      throw new Error('Branch mapping data must have length less than or equal to 2.');
-  }
-
-  return { branchMapping, isRollout, rolloutPercent };
-}
-
-export function logChannelDetails(channel: {
-  branchMapping: string;
-  updateBranches: {
-    updates: (FormatUpdateParameter & {
-      runtimeVersion: string;
-      group: string;
-      platform: string;
-    })[];
-    name: string;
-    id: string;
-  }[];
-}): void {
-  const { branchMapping, isRollout, rolloutPercent } = getBranchMapping(channel.branchMapping);
-
-  const table = new Table({
-    head: ['branch', ...(isRollout ? ['rollout percent'] : []), ...UPDATE_COLUMNS],
-    wordWrap: true,
-  });
-
-  for (const index in branchMapping.data) {
-    if (parseInt(index, 10) > 1) {
-      throw new Error('Branch Mapping data must have length less than or equal to 2.');
-    }
-
-    const { branchId } = branchMapping.data[index];
-    const branch = channel.updateBranches.filter(branch => branch.id === branchId)[0];
-    if (!branch) {
-      throw new Error('Branch mapping is pointing at a missing branch.');
-    }
-    const update = branch.updates[0];
-    table.push([
-      branch.name,
-      ...(isRollout
-        ? [
-            parseInt(index, 10) === 0
-              ? `${rolloutPercent! * 100}%`
-              : `${(1 - rolloutPercent!) * 100}%`,
-          ]
-        : []),
-      formatUpdateMessage(update),
-      update?.runtimeVersion ?? 'N/A',
-      update?.group ?? 'N/A',
-      getPlatformsForGroup({
-        updates: branch.updates,
-        group: branch.updates[0]?.group,
-      }),
-    ]);
-  }
-  Log.log(table.toString());
-}
+import { enableJsonOutput } from '../../utils/json';
 
 export default class ChannelView extends EasCommand {
   static override description = 'view a channel';
@@ -149,17 +23,17 @@ export default class ChannelView extends EasCommand {
   ];
 
   static override flags = {
-    json: Flags.boolean({
-      description: 'print output as a JSON object with the channel ID, name and branch mapping.',
-      default: false,
-    }),
+    ...EasNonInteractiveAndJsonFlags,
+    ...EasPaginatedQueryFlags,
   };
 
   async runAsync(): Promise<void> {
     let {
       args: { name: channelName },
-      flags: { json: jsonFlag },
+      flags,
     } = await this.parse(ChannelView);
+    const paginatedQueryOptions = getPaginatedQueryOptions(flags);
+    const { json: jsonFlag, 'non-interactive': nonInteractive } = flags;
     if (jsonFlag) {
       enableJsonOutput();
     }
@@ -170,39 +44,26 @@ export default class ChannelView extends EasCommand {
 
     if (!channelName) {
       const validationMessage = 'A channel name is required to view a specific channel.';
-      if (jsonFlag) {
+      if (nonInteractive) {
         throw new Error(validationMessage);
       }
-      ({ name: channelName } = await promptAsync({
-        type: 'text',
-        name: 'name',
-        message: 'Provide a channel name:',
-        validate: value => (value ? true : validationMessage),
-      }));
+      const selectedUpdateChannel = await selectChannelOnAppAsync({
+        projectId,
+        selectionPromptTitle: 'Select a channel to view',
+        paginatedQueryOptions: {
+          json: paginatedQueryOptions.json,
+          nonInteractive: paginatedQueryOptions.nonInteractive,
+          offset: 0,
+        },
+      });
+      channelName = selectedUpdateChannel.name;
+      assert(channelName, `A channel must be selected.`);
     }
 
-    const channel = await ChannelQuery.getUpdateChannelByNameForAppAsync({
-      appId: projectId,
+    await listAndRenderBranchesAndUpdatesOnChannelAsync({
+      projectId,
       channelName,
+      paginatedQueryOptions,
     });
-    if (!channel) {
-      throw new Error(`Could not find a channel with name: ${channelName}`);
-    }
-
-    if (jsonFlag) {
-      printJsonOnlyOutput(channel);
-    } else {
-      Log.addNewLineIfNone();
-      Log.log(chalk.bold('Channel:'));
-      Log.log(
-        formatFields([
-          { label: 'Name', value: channel.name },
-          { label: 'ID', value: channel.id },
-        ])
-      );
-      Log.addNewLineIfNone();
-      Log.log(chalk`{bold Branches pointed at this channel and their most recent update group:}`);
-      logChannelDetails(channel);
-    }
   }
 }